requirements.txt
beautifulsoup4
bleach
bunch
certifi
chardet
deepmerge
Django
django-contrib-comments
django-crontab
django-realtime
django-rest-framework
django-rest-auth
django-widget-tweaks
djangorestframework
et-xmlfile
filebrowser-safe
fpdf
future
grappelli-safe
h5py
html5lib
idna
immpload
inflection
isodate
jdcal
joblib
lxml
Markdown
Mezzanine
mysqlclient
numpy
oauthlib
olefile
openpyxl
pandas
parse
pathlib
Pillow
PyMySQL
pyparsing
python-dateutil
python-docx
pytz
PyYAML
rdflib
reportlab
requests
requests-oauthlib
simplejson
six
SPARQLWrapper
tzlocal
Unidecode
urllib3
webencodings
xlwt
